# Python/MonolithicApp/openapi_simplified.yaml
openapi: 3.0.1
info:
  title: City API
  version: 1.0.0
paths:
  /health:
    get:
      summary: Health Check
      description: Health check endpoint
      operationId: health_check_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /city/{name}:
    get:
      summary: Get City
      description: Get city by name
      operationId: get_city_city__name__get
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
            title: Name
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/City'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /city/info/{name}:
    get:
      summary: Get City Info
      description: Get city information
      operationId: get_city_info_city_info__name__get
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
            title: Name
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityInfoDto'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /cities:
    get:
      summary: Get All Cities
      description: Get all cities
      operationId: get_all_cities_cities_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/City'
                type: array
                title: Response Get All Cities Cities Get
  /routes/shortest:
    get:
      summary: Get Shortest Route
      description: Get shortest route between two cities
      operationId: get_shortest_route_routes_shortest_get
      parameters:
        - name: fromName
          in: query
          required: true
          schema:
            type: string
            title: Fromname
        - name: toName
          in: query
          required: true
          schema:
            type: string
            title: Toname
        - name: date
          in: query
          required: false
          schema:
            type: string
            title: Date
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /routes/random-points:
    get:
      summary: Random Points
      description: Generate random points and find shortest route
      operationId: random_points_routes_random_points_get
      parameters:
        - name: count
          in: query
          required: true
          schema:
            type: integer
            title: Count
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                title: Response Random Points Routes Random Points Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /routes/most-used:
    get:
      summary: Get Most Used Connections
      description: Get most used connections
      operationId: get_most_used_connections_routes_most_used_get
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            title: Date
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConnectionUsageDto'
                title: Response Get Most Used Connections Routes Most Used Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'

components:
  schemas:
    City:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        state:
          type: string
          title: State
        population:
          type: integer
          title: Population
      type: object
      required:
        - name
      title: City
    CityInfoDto:
      properties:
        name:
          type: string
          title: Name
        state:
          type: string
          title: State
        population:
          type: integer
          title: Population
        category:
          type: string
          title: Category
        name_length:
          type: integer
          title: Name Length
        vowels:
          type: integer
          title: Vowels
        projection:
          type: integer
          title: Projection
      type: object
      required:
        - name
        - population
        - category
        - name_length
        - vowels
        - projection
      title: CityInfoDto
    ConnectionUsageDto:
      properties:
        connection_id:
          type: string
          title: Connection Id
        from_city:
          type: string
          title: From City
        to_city:
          type: string
          title: To City
        usage_count:
          type: integer
          title: Usage Count
      type: object
      required:
        - connection_id
        - from_city
        - to_city
        - usage_count
      title: ConnectionUsageDto
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HeatmapEdgeUsage:
      properties:
        from_city:
          type: string
          title: From City
        to_city:
          type: string
          title: To City
        count:
          type: integer
          title: Count
      type: object
      required:
        - from_city
        - to_city
        - count
      title: HeatmapEdgeUsage
    HeatmapRequest:
      properties:
        requests:
          items:
            $ref: '#/components/schemas/RequestPair'
          type: array
          title: Requests
        directed:
          type: boolean
          title: Directed
          default: true
        parallelism:
          type: integer
          title: Parallelism
      type: object
      required:
        - requests
      title: HeatmapRequest
    HeatmapResponse:
      properties:
        total_requests:
          type: integer
          title: Total Requests
        edges:
          items:
            $ref: '#/components/schemas/HeatmapEdgeUsage'
          type: array
          title: Edges
        missing:
          items:
            $ref: '#/components/schemas/RequestPair'
          type: array
          title: Missing
      type: object
      required:
        - total_requests
        - edges
        - missing
      title: HeatmapResponse
    RequestPair:
      properties:
        from_city:
          type: string
          title: From City
        to_city:
          type: string
          title: To City
      type: object
      required:
        - from_city
        - to_city
      title: RequestPair
    RouteResponse:
      properties:
        from_city:
          type: string
          title: From City
        to_city:
          type: string
          title: To City
        stops:
          type: integer
          title: Stops
        path:
          items:
            type: string
          type: array
          title: Path
      type: object
      required:
        - from_city
        - to_city
        - stops
        - path
      title: RouteResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError